Move the 2D SIA model to the core modules
